Update to new package structure
