+++ about/src/main/java/io/plaidapp/about/ui/adapter/LibraryIntroHolder.kt
@@ -14,13 +14,9 @@
  * limitations under the License.
  */
 
-package io.plaidapp.ui.about.uimodel
+package io.plaidapp.about.ui.adapter
 
-/**
- * Hold values displayed in the about Ui.
- */
-internal data class AboutUiModel(
-    val appAboutText: CharSequence,
-    val iconAboutText: CharSequence,
-    val librariesUiModel: LibrariesUiModel
-)
+import android.support.v7.widget.RecyclerView
+import android.view.View
+
+internal class LibraryIntroHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
